requirements.txt
requests
deep_translator
python-dotenv
transformers
opencv-python
mediapipe
fer
dlib
pandas
moviepy
